buildings: fixed building ruins

+++ src/building/list.h
@@ -58,4 +58,4 @@ void building_list_burning_clear(void);
 
 void building_list_burning_add(int building_id);
 
-std::span<int> building_list_burning_items();
+std::span<uint16_t> building_list_burning_items();
